+++ Addons/PolyQuilt/subtools/maintool_delete.py
@@ -24,7 +24,6 @@ from ..QMesh import *
 from ..utils.mouse_event_util import ButtonEventUtil, MBEventType
 from .subtool import *
 from .subtool_knife import *
-from .subtool_edge_slice import *
 from .subtool_edgeloop_cut import *
 from .subtool_delete import *
