+++ Addons/PolyQuilt/subtools/subtool_edgeloop_extrude.py
@@ -24,54 +24,55 @@ from ..utils import pqutil
 from ..utils import draw_util
 from ..QMesh import *
 from ..utils.dpi import *
-from .subtool import SubTool
+from .subtool import MainTool
 
-class SubToolEdgeLoopExtrude(SubTool) :
-    name = "EdgeLoopExtrudeTool"
+class SubToolEdgeLoopExtrude(MainTool) :
+    name = "EdgeLoop Extrude"
 
-    def __init__(self,op, target : ElementItem , is_loop : bool = True ) :
-        super().__init__(op)
+    def __init__(self,op,target, button) :
+        super().__init__(op,target, button , no_hold = True )      
         self.l2w = self.bmo.local_to_world_pos
         self.w2l = self.bmo.world_to_local_pos
 
         self.currentVert = None
         self.currentEdge = None
         self.moveType = 'NORMAL'
-        if target.isEdge :
-            self.currentEdge = target.element
-            if is_loop :
-                self.edges , __ = self.bmo.findEdgeLoop( self.currentEdge  )
-            else :
-                self.edges = [ self.currentEdge ]
 
-            self.verts = set()
-            for e in self.edges :
-                self.verts = self.verts | set( e.verts )
-        elif target.isVert :
-            self.moveType = 'SPREAD'
-            self.currentVert = target.element
-            self.edges , self.verts = self.bmo.findOutSideLoop( self.currentVert )
+        self.currentEdge = target.element
+        self.edges , __ = self.bmo.calc_edge_loop( self.currentEdge , is_mirror = False )
+
+        self.verts = set()
+        for e in self.edges :
+            self.verts = self.verts | set( e.verts )
 
         self.verts = { v : self.l2w( v.co ) for v in self.verts }
 
+        self.mirror_edge = None
+
         # mirror
+        is_symetry = False
+        self.mirrorEdges = []
         if self.bmo.is_mirror_mode :
-            self.centerVerts = [ v for v in self.verts if self.bmo.is_x_zero_pos(v.co) ]
-            if target.hitPosition.x >= 0 :
-                self.plusVerts = { v : self.bmo.find_mirror( v ) for v in self.verts if v.co.x > 0 and v not in self.centerVerts }
-                self.minusVerts = { v : self.bmo.find_mirror( v ) for v in self.verts if v not in self.plusVerts and v not in self.centerVerts }
+            mirror_edges = { e : self.bmo.find_mirror( e ) for e in self.edges }
+            if set( self.edges ) & set(mirror_edges.values()) :
+                self.centerVerts = [ v for v in self.verts if self.bmo.is_x_zero_pos(v.co) ]
+                notCenterVerts = [ v for v in self.verts if v not in self.centerVerts ]
+                if target.hitPosition.x >= 0 :
+                    self.plusVerts = [ v for v in notCenterVerts if v.co.x > 0 ]
+                else :
+                    self.plusVerts = [ v for v in notCenterVerts if v.co.x < 0 ]
+                self.minusVerts = { self.bmo.find_mirror( v ) : v  for v in notCenterVerts if v not in self.plusVerts }
+                is_symetry = True
             else :
-                self.plusVerts = { v : self.bmo.find_mirror( v ) for v in self.verts if v.co.x < 0 and v not in self.centerVerts }
-                self.minusVerts = { v : self.bmo.find_mirror( v ) for v in self.verts if v not in self.plusVerts and v not in self.centerVerts }
-            self.mirrorEdges = [ e for e in self.edges if self.bmo.find_mirror( e ) in self.edges ]
-        else :
+                self.mirrorEdges = { v : m for v , m in mirror_edges.items() if m }
+
+        if not is_symetry :
             if self.preferences.fix_to_x_zero :
                 self.centerVerts = [ v for v in self.verts if self.bmo.is_x_zero_pos(v.co) ]
             else :
                 self.centerVerts = []
             self.plusVerts = {v : None for v in self.verts if v not in self.centerVerts }
             self.minusVerts = {}
-            self.mirrorEdges = []
 
         self.startPos = target.hitPosition
         self.targetPos = target.hitPosition
@@ -90,11 +91,33 @@ class SubToolEdgeLoopExtrude(SubTool) :
 
     @staticmethod
     def Check( root , target ) :
-        if target.isVert :
-            return target.element.is_boundary 
-        elif target.isEdge :
-            return target.element.is_boundary or not target.element.is_manifold
+        if target.isEdge :
+            return target.element.is_boundary
         return False
+
+    @staticmethod
+    def pick_element( qmesh , location , preferences ) :
+        element = qmesh.PickElement( location , preferences.distance_to_highlight , edgering = True , elements = ["EDGE"] )
+        if element.isEdge : 
+            if not element.element.is_convex :
+                return ElementItem.Empty()
+        return element        
+
+    @classmethod
+    def DrawHighlight( cls , gizmo , element ) :
+        lw2 = gizmo.bmo.local_to_world_pos
+        if element.isEdge :
+            loop_edge , v = gizmo.bmo.calc_edge_loop( element.element )        
+            color = gizmo.preferences.makepoly_color         
+            draw_lines = ( ( lw2( e.verts[0].co ) , lw2( e.verts[1].co) ) for e in loop_edge )
+            draw_lines = sum( draw_lines , () )
+            def draw() :
+                draw_util.draw_lines3D( bpy.context , draw_lines , color , 4 , primitiveType = 'LINES' , hide_alpha = 0 )
+            return draw
+
+        return None
+
+
 
     def OnUpdate( self , context , event ) :
         if event.type == 'MOUSEMOVE':
@@ -104,22 +127,23 @@ class SubToolEdgeLoopExtrude(SubTool) :
             vG = self.move_plane.intersect_ray( rayG )
             move = (vG - vS)
             self.targetPos = self.startPos + move
+            if QSnap.is_active() :
+                self.targetPos = QSnap.view_adjust( self.targetPos )
             dist = self.preferences.distance_to_highlight
 
             self.is_center_snap = False
             if self.bmo.is_mirror_mode :
                 self.is_center_snap = self.bmo.is_x0_snap( self.targetPos )
 
-            def adjustVert( vt , zero_snap ) :
-                if self.is_center_snap :
-                    vt.x = 0
-
-                pt = QSnap.adjust_point( vt )
+            def adjustVert( v , co , zero_snap ) :
+                if self.is_center_snap or zero_snap :
+                    co = self.bmo.zero_pos_w2l(co)
+                pt = QSnap.adjust_point( co )
                 ct = pqutil.location_3d_to_region_2d(pt)
                 elm = self.bmo.PickElement( ct , dist , edgering=True , backface_culling = True , elements=['VERT'], ignore=self.ignoreVerts )
                 if elm.isVert :
                     return elm.element
-                if zero_snap :
+                if self.is_center_snap or zero_snap :
                     pt = self.bmo.zero_pos_w2l( pt )
                 return pt
 
@@ -151,27 +175,15 @@ class SubToolEdgeLoopExtrude(SubTool) :
             # 各頂点の移動
             for v in self.centerVerts :
                 p = calcMove( v , move )
-                self.verts[v] = adjustVert(p,True)
+                self.verts[v] = adjustVert(v,p,True)
 
             for v in self.plusVerts :
                 p = calcMove( v , move )
-                self.verts[v] = adjustVert(p,False)
+                self.verts[v] = adjustVert(v,p,False)
 
-            for v in self.minusVerts :
-                r = self.minusVerts[v]
-                if r != None :
-                    s = self.verts[r]
-                    if isinstance( s , bmesh.types.BMVert ) :
-                        t = self.bmo.find_mirror( s )
-                        if t == None :
-                            self.verts[v] = self.bmo.mirror_pos(self.l2w(s.co))
-                        else :
-                            self.verts[v] = t
-                    else :
-                        self.verts[v] = self.bmo.mirror_pos_w2l(s)
-                else :
-                    p = self.bmo.mirror_pos_w2l(self.l2w(v.co) + move)
-                    self.verts[v] = adjustVert(p,False)
+            for v,m in self.minusVerts.items() :
+                p = calcMove( m , mathutils.Vector(( -move.x , move.y , move.z ))  )
+                self.verts[m] = adjustVert(v,p,False)
 
             # スナップする辺を探す
             self.snapTarget = ElementItem.Empty()
@@ -180,12 +192,27 @@ class SubToolEdgeLoopExtrude(SubTool) :
                 if snapTarget.isEdge :
                     self.snapTarget = snapTarget
                     self.AdsorptionEdge( self.currentEdge , snapTarget.element )
+                    # スナップ結果から左右の対称補正
+                    for v in self.centerVerts :
+                        if isinstance( self.verts[v] , bmesh.types.BMVert ) :
+                            if not self.bmo.is_x_zero_pos_w2l( self.verts[v].co ) :
+                                p = calcMove( v , move )
+                                self.verts[v] = adjustVert(v,p,True)
+
+                    for v,m in self.minusVerts.items() :
+                        if  m != None and v != None :
+                            if isinstance( self.verts[v] , bmesh.types.BMVert ) :
+                                x = self.bmo.find_mirror( self.verts[v] )
+                                if x != None :
+                                    self.verts[m] = x
+                            else :
+                                self.verts[m] = self.bmo.mirror_pos_w2l( self.verts[v] )
 
         elif event.type == 'RIGHTMOUSE' :
             if event.value == 'PRESS' :
                 pass
             elif event.value == 'RELEASE' :
-                pass
+                return 'FINISHED'
         elif event.type == 'LEFTMOUSE' :
             if event.value == 'RELEASE' :
                 self.MakePoly()
@@ -210,51 +237,47 @@ class SubToolEdgeLoopExtrude(SubTool) :
             draw_util.draw_circle2D( pos , size , (1,1,1,1) , False )
 
     def OnDraw3D( self , context  ) :
-        def v2p( e , v ) :
-            if isinstance( v , mathutils.Vector ) :
-                return v
-            elif isinstance( v , bmesh.types.BMVert ) :
-                if v not in e.verts :
+        def v2p( e , v , mirror ) :
+            if mirror :
+                if isinstance( v , mathutils.Vector ) :
+                    return self.bmo.mirror_pos_w2l( v )
+                elif isinstance( v , bmesh.types.BMVert ) and v not in e.verts :
+                    return self.l2w( self.bmo.mirror_pos( v.co ) )
+            else :
+                if isinstance( v , mathutils.Vector ) :
+                    return v
+                elif isinstance( v , bmesh.types.BMVert ) and v not in e.verts :
                     return self.l2w( v.co )
             return None
 
         for e in self.edges :
             p = [ self.l2w( v.co ) for v in e.verts ]
-            t = [ v2p(e, self.verts[v] ) for v in e.verts ]
-            polys = [ v for v in (p[0],t[0],t[1],p[1]) if v != None ]
-            draw_util.draw_Poly3D( self.bmo.obj , polys , self.color_create(0.5), hide_alpha = 0.5  )        
-            draw_util.draw_lines3D( context , polys , self.color_create(1.0) , 2 , primitiveType = 'LINE_LOOP' , hide_alpha = 0 )        
-            if self.snapTarget.isEdge and None not in t :
-                draw_util.draw_lines3D( context , [ t[0] , t[1] ] , (1,1,1,1) , 3 , primitiveType = 'LINE_STRIP' , hide_alpha = 1 )
+            t = [ v2p(e, self.verts[v],False ) for v in e.verts ]
+            polyss = [ [ v for v in (p[0],t[0],t[1],p[1]) if v != None ] ]
 
-            if self.bmo.is_mirror_mode and e not in self.mirrorEdges :
-                polys = [ self.bmo.mirror_pos_w2l(p) for p in polys ]
-                draw_util.draw_Poly3D( self.bmo.obj , polys , self.color_create(0.25), hide_alpha = 0.5  )        
-                draw_util.draw_lines3D( context , polys , self.color_create(0.5) , 1 , primitiveType = 'LINE_STRIP' , hide_alpha = 0 )        
-                if self.is_center_snap :            
-                    draw_util.draw_lines3D( context , [ t[0] , t[1] ] , (1,1,1,1) , 2 , primitiveType = 'LINE_STRIP' , hide_alpha = 1 )
+            if self.mirrorEdges and e in self.mirrorEdges :
+                polyss.append( [ self.bmo.mirror_pos_w2l(v) for v in (p[0],t[0],t[1],p[1]) if v != None ] )
+
+            for polys in polyss :
+                draw_util.draw_Poly3D( self.bmo.obj , polys , self.color_create(0.5), hide_alpha = 0.5  )        
+                draw_util.draw_lines3D( context , polys , self.color_create(1.0) , 2 , primitiveType = 'LINE_LOOP' , hide_alpha = 0 )        
+                if self.snapTarget.isEdge and None not in t :
+                    draw_util.draw_lines3D( context , [ t[0] , t[1] ] , (1,1,1,1) , 3 , primitiveType = 'LINE_STRIP' , hide_alpha = 1 )
 
     def AdsorptionEdge( self , srcEdge , snapEdge ) :
-        def v2p( v ) :
-            if isinstance( v , bmesh.types.BMVert ) :
-                return self.l2w( v.co )
-            return v
+        dstEdges , __ = self.bmo.calc_edge_loop( snapEdge )        
 
-        p0 = v2p( self.l2w(srcEdge.verts[0].co) )
-        p1 = v2p( self.l2w(srcEdge.verts[1].co) )
+        p0 = self.l2w( srcEdge.verts[0].co)
+        p1 = self.l2w( srcEdge.verts[1].co) 
+        s0 = self.l2w( snapEdge.verts[0].co) 
+        s1 = self.l2w( snapEdge.verts[1].co) 
 
-        st0 = pqutil.location_3d_to_region_2d(p0)
-        st1 = pqutil.location_3d_to_region_2d(p1)
-        se0 = pqutil.location_3d_to_region_2d(self.bmo.local_to_world_pos(srcEdge.verts[0].co))
-        se1 = pqutil.location_3d_to_region_2d(self.bmo.local_to_world_pos(srcEdge.verts[1].co))
-        if (st0-se0).length + (st1-se1).length > (st0-se1).length + (st1-se0).length :
-            t0 = snapEdge.verts[1]
-            t1 = snapEdge.verts[0]
-        else :
+        if (p0-s0).length + (p1-s1).length > (p0-s1).length + (p1-s0).length :
             t0 = snapEdge.verts[0]
             t1 = snapEdge.verts[1]
-
-        dstEdges = self.bmo.findOutSideEdgeLoop( snapEdge , t0 )        
+        else :
+            t0 = snapEdge.verts[1]
+            t1 = snapEdge.verts[0]
 
         def other( edge , vert , edges ) :
             hits = [ e for e in edges if e != edge and vert in e.verts ]
@@ -268,19 +291,22 @@ class SubToolEdgeLoopExtrude(SubTool) :
             dv = dst
             de = snapEdge
             while( sv != None and dv != None ) :
-                if sv != None and dv != None :
-                    self.verts[sv] = dv
+                self.verts[sv] = dv
                 se , sv = other( se , sv , self.edges  )
                 de , dv = other( de , dv , dstEdges  )
 
     def MakePoly( self ) :
-        mirror = None
+        threshold = bpy.context.scene.tool_settings.double_threshold
+
+        if (self.targetPos - self.startPos  ).length <= threshold :
+            return
 
         for vert in self.verts :
             if isinstance( self.verts[vert] , mathutils.Vector ) :
                 self.verts[vert] = self.bmo.AddVertexWorld( self.verts[vert] , False )
                 self.bmo.UpdateMesh()
 
+        newFaces = []
         for edge in self.edges :
             t = [ self.verts[v] for v in edge.verts ]
             if  t[0] == None and t[1] == None :
@@ -295,10 +321,12 @@ class SubToolEdgeLoopExtrude(SubTool) :
                             verts.reverse()
             else :
                 normal = pqutil.getViewDir()
-            if edge in self.mirrorEdges :
-                mirror = False
-            else:
-                mirror = None
 
-            face = self.bmo.AddFace( verts , normal , mirror )
+            newFaces.append( self.bmo.AddFace( verts , normal , is_mirror = (len(self.mirrorEdges) > 0 ) ) )
             self.bmo.UpdateMesh()
+
+        newVerts = set( sum( ( tuple(f.verts) for f in newFaces ) , () ) )
+
+        bmesh.ops.remove_doubles( self.bmo.bm , verts = list(newVerts) , dist = threshold )
+
+        self.bmo.UpdateMesh()
